requirements.txt
absl-py==2.0.0
astunparse==1.6.3
bidict==0.22.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.0.3
Flask-Cors==4.0.0
Flask-SocketIO==4.3.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.0.3
wrapt==1.15.0
wsproto==1.2.0